Added Video Filter project, adjust projects page UI
Added two new projects, the Video Filter and the Video Filter with Cpp front end.

Additionally, the Large View now displays above the Small View project tiles, this makes the screen feel more full. In order to accomodate for mobile, the screen now scrolls to the top of the site upon opening a large view.